[FLINK-17547][task][hotfix] Move RefCountedFile to flink-core to use it in SpanningWrapper
