+++ flink-core/src/main/java/org/apache/flink/util/RefCounted.java
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package org.apache.flink.fs.s3.common.utils;
+package org.apache.flink.util;
 
 import org.apache.flink.annotation.Internal;
